+++ packages/react/src/Suspense.tsx
@@ -1,7 +1,7 @@
 import { type ComponentProps, type ComponentType, type SuspenseProps as ReactSuspenseProps, useContext } from 'react'
-import { SuspenseDefaultPropsContext } from './contexts'
-import type { PropsWithoutChildren } from './utility-types'
-import { defineSuspense } from './utils'
+import { SuspenseDefaultPropsContext } from './contexts/DefaultPropsContexts'
+import type { PropsWithoutChildren } from './utility-types/PropsWithoutChildren'
+import { defineSuspense } from './utils/defineSuspense'
 
 export interface SuspenseProps extends ReactSuspenseProps {
   /**
@@ -16,15 +16,16 @@ export interface SuspenseProps extends ReactSuspenseProps {
  * @see {@link https://suspensive.org/docs/react/Suspense Suspensive Docs}
  */
 export const Suspense = Object.assign(
-  ({ clientOnly, children, fallback }: SuspenseProps) => {
+  ({ clientOnly, fallback, children, name, ...rest }: SuspenseProps) => {
     const defaultProps = useContext(SuspenseDefaultPropsContext)
     const DefinedSuspense = defineSuspense({
       defaultPropsClientOnly: defaultProps.clientOnly,
       componentPropsClientOnly: clientOnly,
     })
-
     return (
-      <DefinedSuspense fallback={fallback === undefined ? defaultProps.fallback : fallback}>{children}</DefinedSuspense>
+      <DefinedSuspense {...rest} fallback={fallback === undefined ? defaultProps.fallback : fallback} name={name}>
+        {children}
+      </DefinedSuspense>
     )
   },
   {