+++ packages/react/src/contexts/index.ts
@@ -1,1 +1,0 @@
-export { DelayDefaultPropsContext, SuspenseDefaultPropsContext } from './DefaultPropsContexts'