+++ packages/react-query-5/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @suspensive/react-query-5
 
+## 3.3.2
+
+## 3.3.1
+
 ## 3.3.0
 
 ### Minor Changes