+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @suspensive/react
 
+## 3.3.2
+
+### Patch Changes
+
+- [`c4c0bf8`](https://github.com/toss/suspensive/commit/c4c0bf8ae543d43ba4fd61a59565bedfdcf7452b) Thanks [@manudeli](https://github.com/manudeli)! - fix(react): enable additional props and add 'name' prop to Suspense component (#1663)
+
+## 3.3.1
+
+### Patch Changes
+
+- [#1661](https://github.com/toss/suspensive/pull/1661) [`27bda1d`](https://github.com/toss/suspensive/commit/27bda1df5675c068e2cbb6f202b4ceb00ced4465) Thanks [@manudeli](https://github.com/manudeli)! - perf(react, react-query): reduce size
+
 ## 3.3.0
 
 ## 3.2.3